requirements.txt
# AI Coding Workflow Manager - Requirements

# Core dependencies
anthropic>=0.18.0
openai>=1.0.0
python-dotenv>=1.0.0

# Async support
aiohttp>=3.9.0
asyncio>=3.4.3

# Git integration
gitpython>=3.1.40

# Data processing
pydantic>=2.0.0
pyyaml>=6.0.0

# CLI tools
click>=8.1.0
rich>=13.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
